+++ lib/mongoid/fields/internal/true_class.rb
@@ -1,0 +1,10 @@
+# encoding: utf-8
+module Mongoid #:nodoc:
+  module Fields #:nodoc:
+    module Internal #:nodoc:
+      # Defines the behaviour for true values.
+      class TrueClass < Boolean
+      end
+    end
+  end
+end